modifs ranked

+++ LolStatistics.Web/Models/GameViewModel.cs
@@ -11,13 +11,13 @@ namespace LolStatistics.Web.Models
 
         public int Spell2 { get; set; }
 
-        public int Kills { get; set; }
+        public long Kills { get; set; }
 
-        public int Assists { get; set; }
+        public long Assists { get; set; }
 
-        public int Deaths { get; set; }
+        public long Deaths { get; set; }
 
-        public int[] Items { get; set; }
+        public long[] Items { get; set; }
 
         public int GameDuration { get; set; }
